aggregator-service/requirements.txt
flask
mysql-connector-python
sqlalchemy
psycopg2-binary
